Change browser key to main
Was seeing failed tests in watchtower due to not being able to find
this package. The "browser" key was having no effect in the package,
but it seems to be meant to serve the same purpose as the "main" key
(primary entry point to the program). Changed to main key and tests
are passing.